feat: modularizes popup as a promise
