+++ js/btm.js
@@ -4,13 +4,12 @@ import { spectrumSites, siteTitles, getSlug, createPopup, siteSearches, getPubli
 
 import { getLinks, checkIsArticleHead, checkLinkSection, checkIsProperSource } from './helpers/getLinks-helpers'
 
-import { getPopoverHtml, getBTMIcon, getLoading, getPopoverTitle, getArticlePagePopover } from './helpers/inline-elements'
+import { getBTMIcon, getLoading, getArticlePagePopover } from './helpers/inline-elements'
 
-import { getPopoverSide, reposition, toggleSummary, openArticleLink } from './helpers/embed-helpers'
+import { getPopoverSide, toggleSummary, openArticleLink, placePopover } from './helpers/embed-helpers'
 
 $(() => {
 	const domain = window.location.hostname.split('www.')[1]
-		, originTitle = siteTitles[domain] || domain
 		, pathname = window.location.pathname
 		, source = siteTitles[domain] || domain
 		, btmIcon = chrome.runtime.getURL('assets/btm_logo.png')
@@ -22,10 +21,9 @@ $(() => {
 		, startTime = new Date(); // this is initialized at the current time
 
 	function checkFacebookLinks() {
-		/* eslint no-prototype-builtins: "error" */
 		let $links = $('a').toArray().filter(link => link.href && !Object.prototype.hasOwnProperty.call(hrefs, link.href) && checkIsProperSource(link) && checkIsArticleHead(link) && checkLinkSection(link.href))
 
-		$links = uniqBy($links, link => link.href) // filters out duplicates
+		$links = uniqBy($links, link => link.href)
 
 		$links.forEach(element => {
 			hrefs[element.href] = true;
@@ -37,41 +35,14 @@ $(() => {
 				, $postText = $newsfeedPost.find('.userContent')
 				, $btmButton = getBTMIcon(btmIcon, slug)
 				, publisher = getPublisher(href)
-			let side = "right";
+				, placeButton = new Promise((resolve, reject) => {
+					$postText.first().append($btmButton) ? resolve($btmButton.offset()) : reject('Not placed')
+				})
 
-			$btmButton.popover({
-				trigger: "click",
-				container: "body",
-				html: "true",
-				template: getPopoverHtml(slug),
-				placement: (popover, parent) => {
-					const distFromRight = $(window).width() - $(parent).offset().left
-					side = (distFromRight < 350) ? "left" : "right";
-					return side;
-				},
-				title: getPopoverTitle(btmBg, btmIcon),
-				content: getLoading(slug)
-			})
-
-			$postText.first().append($btmButton);
-
-			function initPopover() {
-				$('.btm-close').on('click', () => { $btmButton.popover('hide') });
-				Promise.all(siteSearches(spectrumSites[publisher], slug))
-					.then(results => {
-						$(`#btm-loading-${slug}`).hide();
-						$(`#btm-popover-body-${slug}`).after(createPopup(results, slug));
-						reposition(slug, side);
-						$('.collapse-link').on('click', toggleSummary);
-						$('.popup-link').on('click', (event) => openArticleLink(event, 'Facebook'));
-					})
-				chrome.runtime.sendMessage({
-					source: originTitle,
-					type: "BTM Icon Click"
-				});
-			}
-
-			$btmButton.on('shown.bs.popover', () => initPopover(slug));
+			placeButton
+				.then(offset => getPopoverSide(offset))
+				.then(side => placePopover(side, $btmButton, slug, btmBg, btmIcon, source, publisher, startTime))
+				.catch(console.error)
 		})
 	}
 
@@ -119,49 +90,32 @@ $(() => {
 	function embedIcons() {
 		const $links = getLinks()
 
-		$links.forEach(element => { // this is for nytimes only. not general
+		$links.forEach(element => {
 			const $element = $(element)
 				, href = $element.attr('href')
 				, slug = getSlug(href)
-				, $btmButton = getBTMIcon(btmIcon, slug);
+				, $btmButton = getBTMIcon(btmIcon, slug)
+				, publisher = getPublisher(href)
+				, placeButton = new Promise((resolve, reject) => {
+					if ($element.find('h2.headline a').toArray().length === 0) {
+						if ($element.find('h2.headline').toArray().length > 0) {
+							$btmButton.appendTo($element.find('h2.headline').toArray()[0])
+							resolve($btmButton.offset())
+						} else if (!$element.next().is('a') && $element.attr('class') !== 'popup-link') {
+							$btmButton.insertAfter($element);
+							resolve($btmButton.offset())
+						} else {
+							reject('Not placeable')
+						}
+					} else {
+						reject('Not placeable')
+					}
+				})
 
-			if ($element.find('h2.headline a').toArray().length === 0) {
-				if ($element.find('h2.headline').toArray().length > 0) {
-					$btmButton.appendTo($element.find('h2.headline').toArray()[0])
-				} else if (!$element.next().is('a') && $element.attr('class') !== 'popup-link') {
-					$btmButton.insertAfter($element);
-				}
-			}
-
-			const side = getPopoverSide(slug)
-
-			$btmButton.popover({
-				trigger: "click",
-				container: "body",
-				html: "true",
-				template: getPopoverHtml(slug, side),
-				title: getPopoverTitle(btmBg, btmIcon),
-				placement: side,
-				content: getLoading(slug)
-			})
-
-			function initPopover() {
-				$('.btm-close').on('click', () => { $btmButton.popover('hide') });
-				Promise.all(siteSearches(spectrumSites[domain], slug))
-					.then(results => { // this is the promise part of the site
-						$(`#btm-loading-${slug}`).hide();
-						$(`#btm-popover-body-${slug}`).after(createPopup(results, slug));
-						reposition(slug, side);
-						$('.collapse-link').on('click', toggleSummary);
-						$('.popup-link').on('click', (event) => openArticleLink(event, source, startTime));
-					})
-				chrome.runtime.sendMessage({
-					source,
-					type: "BTM Icon Click"
-				});
-			}
-
-			$btmButton.on('shown.bs.popover', () => initPopover(slug, href));
+			placeButton
+				.then(offset => getPopoverSide(offset))
+				.then(side => placePopover(side, $btmButton, slug, btmBg, btmIcon, source, publisher, startTime))
+				.catch(console.error)
 		})
 	}
