+++ js/helpers/embed-helpers.js
@@ -1,5 +1,7 @@
-export const getPopoverSide = (slug) => {
-	const iconOffset = $(`#btm-icon-${slug}`).offset();
+import { getPopoverHtml, getPopoverTitle, getLoading } from './inline-elements'
+import { siteSearches, spectrumSites, createPopup } from './site-constants'
+
+export const getPopoverSide = (iconOffset) => {
 	const distFromRight = $(window).width() - iconOffset.left
 	return (distFromRight < 350) ? "left" : "right"
 }
@@ -16,6 +18,7 @@ export const reposition = (slug, side) => {
 	}
 	$(`[data-slug='${slug}']`).offset(newOffset)
 	$(`[data-slug='${slug}'].popover-content`).css("top", "0")
+	$(`.btm-arrow.loading`).css("display", "inline-block")
 }
 
 export const openArticleLink = (event, source, startTime) => {
@@ -48,3 +51,30 @@ export const toggleSummary = (event) => {
 	}
 }
 
+export const placePopover = (side, $btmButton, slug, btmBg, btmIcon, source, publisher, startTime) => {
+	$btmButton.popover({
+		trigger: "click",
+		container: "body",
+		html: "true",
+		template: getPopoverHtml(slug, side),
+		placement: side,
+		title: getPopoverTitle(btmBg, btmIcon),
+		content: getLoading(slug)
+	})
+	function initPopover() {
+		$('.btm-close').on('click', () => { $btmButton.popover('hide') });
+		Promise.all(siteSearches(spectrumSites[publisher], slug))
+			.then(results => {
+				$(`#btm-loading-${slug}`).hide();
+				$(`#btm-popover-body-${slug}`).after(createPopup(results, slug));
+				reposition(slug, side);
+				$('.collapse-link').on('click', toggleSummary);
+				$('.popup-link').on('click', (event) => openArticleLink(event, source, startTime));
+			})
+		chrome.runtime.sendMessage({
+			source,
+			type: "BTM Icon Click"
+		});
+	}
+	$btmButton.on('shown.bs.popover', () => initPopover());
+}
