+++ stylesheets/index.scss
@@ -73,7 +73,7 @@ $box-shadow: 0 5px 10px rgba(0, 0, 0, .2);
 }
 
 .btm-popover-body {
-  padding: 0 1em;
+  padding: 0 1em 1em 1em;
   font-family: 'Helvetica Neue', sans-serif;
 }
